+++ worker/src/Channel/ChannelRequest.cpp
@@ -37,6 +37,7 @@ namespace Channel
 		{ "transport.connect",                           ChannelRequest::MethodId::TRANSPORT_CONNECT                                },
 		{ "transport.setMaxIncomingBitrate",             ChannelRequest::MethodId::TRANSPORT_SET_MAX_INCOMING_BITRATE               },
 		{ "transport.setMaxOutgoingBitrate",             ChannelRequest::MethodId::TRANSPORT_SET_MAX_OUTGOING_BITRATE               },
+		{ "transport.setMinOutgoingBitrate",             ChannelRequest::MethodId::TRANSPORT_SET_MIN_OUTGOING_BITRATE               },
 		{ "transport.restartIce",                        ChannelRequest::MethodId::TRANSPORT_RESTART_ICE                            },
 		{ "transport.produce",                           ChannelRequest::MethodId::TRANSPORT_PRODUCE                                },
 		{ "transport.consume",                           ChannelRequest::MethodId::TRANSPORT_CONSUME                                },