+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 ### NEXT
 
+* Add `transport.setMinOutgoingBitrate()` method ([PR #1038](https://github.com/versatica/mediasoup/pull/1038), credits to @ jcague).
 * `RTC::RetransmissionBuffer`: Increase `RetransmissionBufferMaxItems` from 2500 to 5000.